requirements.txt
Scrapy==2.5.0
psycopg2_binary>=2.8
requests==2.25.1
dataclasses==0.8
